be/app/uber/pyproject.toml
[tool.poetry]
name = "uber"
version = "0.1.0"
description = ""
authors = ["Gshopping <zays.kim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"
pymongo = "^4.11.1"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
